Implement Web Storage API and refactor
All of the user's tasks and todo's are now saved locally.

Functions such as updateProject() and getProject() in project.js update
and retrieve projects when site is loaded or when changes are made.

list.js now adds methods to each task object when needed so tasks can be
added, removed, and edited.